refactor: #17 헤더 리팩토링
