+++ hooks/useFirstLaunch.ts
@@ -3,7 +3,6 @@ import AsyncStorage from '@react-native-async-storage/async-storage';
 
 export default function Ap() {
   const [firstLaunch, setFirstLaunch] = useState(Boolean);
-
   useEffect(() => {
     AsyncStorage.getItem('launched').then((value) => {
       if (value == null) {
